Commit: Pending testing for Heatmap Logic Decoupling
