+++ server/index.js
@@ -10,7 +10,8 @@ const tvRoutes = require("./routes/tvRoutes");
 const authRoutes = require("./routes/authRoutes");
 const adsRoutes = require("./routes/adsRoutes");
 const roleRoutes = require("./routes/roleRoutes");
-const heatmapRoutes = require("./routes/heatmapRoutes");
+const adAnalyticsRoutes = require("./routes/adAnalyticsRoutes");
+const adAggregatesRoutes = require("./routes/adAggregateRoutes");
 const dashboardRoutes = require("./routes/dashboardRoutes");
 const { handleWebSocketMessage } = require("./controllers/webSocketController");
 
@@ -51,7 +52,8 @@ app.use("/api", authRoutes);
 app.post("/generate-presigned-url", generatePresignedUrlController);
 app.use('/api/ads', adsRoutes);
 app.use('/api/roles', roleRoutes);
-app.use("/api/heatmap", heatmapRoutes);
+app.use("/api/adAnalytics", adAnalyticsRoutes);
+app.use("/api/adAggregates", adAggregatesRoutes);
 app.use("/api/dashboard", dashboardRoutes);
 
 // WebSocket Server Handling
